requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
gitpython==3.1.41
python-dotenv==1.0.0 